BackEnd_Equipo3/requirements.txt
pandas
kafka-python
<<<<<<< HEAD
pymongo
psycopg2-binary
=======
>>>>>>> dev
